Bug 1644610 - Add ffmpeg/VA-API decoding for X11 Basic compositor.
Add a presentation path using X Pixmaps. Decoder renders a decoded image
into an X Pixmap, which is then passed to the rendering side. Although
this approach can be used for basic, opengl and WebRender, the path is
currently enabled only for the basic compositor. Once required bits of
X Pixmap-based TextureHosts are implemented for opengl and WebRender
compositors, they can also utilize this path.

Differential Revision: https://phabricator.services.mozilla.com/D79012